Switch to minitest for consistent output between Ruby versions
In Ruby 2.0, `test/unit` activates a fancy output style for terminals
that doesn't look well in Travis logs. I could disable it with the
`--jobs-status=normal` test optiont, but that option isn't supported in
Ruby 1.8.7.

Instead of dealing with inconsistencies between test support in
different Ruby versions, I'm simply switching to a fixed version of
Minitest.