Reduce features in SshConfig scanner for speed
A lot of hub execution time was spent in SshConfig, which was a
reasonably compatible reader implementation for the format described in
ssh_config(5). However, all we ever used this functionality for is
resolving hostname aliases the same way `ssh` on the command-line does.
See 4b57f61e5af42f47b79dc2522a595d66817aa0f7

To speed up hub, this strips SshConfig feature set down to a bare
minimum to keep the original functionality:

- Don't record all configuration keys, just "HostName"
- Drop support for the "key=value" assignment syntax
- Drop support for hostname patterns, including "*" to match any value